Making parser singlethreaded

+++ parser/src/ast/helper.rs
@@ -76,12 +76,13 @@ impl FunctionCall {
         }
     }
 }
-
+///combines the dependants into one vec
 trait SimpleDependantDrain {
-    fn concat_dependant(&self) -> Vec<u32>;
+    ///drain dependants to one vec
+    fn concat_dependant(&self) -> Vec<u64>;
 }
 impl<U: Simplify, V: Simplify> SimpleDependantDrain for (Simple<U>, Simple<V>) {
-    fn concat_dependant(&self) -> Vec<u32> {
+    fn concat_dependant(&self) -> Vec<u64> {
         [&self.0.dependents[..], &self.1.dependents[..]].concat()
     }
 }
@@ -242,18 +243,22 @@ impl SimpleCompareMultipleMathExprs for (Simple<MathExpr>, Simple<MathExpr>) {
 /// MathExpr is in the simplest form
 #[derive(Clone, Debug, PartialEq, Hash)]
 pub struct Simple<T: Simplify> {
+    ///The value that is simplified
     pub(crate) value: T,
-    dependents: Vec<u32>,
+    ///The dependants of the simplified values
+    dependents: Vec<u64>,
 }
 impl<T: Simplify> Simple<T> {
     ///returns the inner T in a box
     pub fn boxed_inner(self) -> Box<T> {
         Box::new(self.inner())
     }
-    pub const fn dependant(&self) -> &Vec<u32> {
+    /// Gets the dependants
+    pub const fn dependant(&self) -> &Vec<u64> {
         &self.dependents
     }
-    pub fn destruct(self) -> (T, Vec<u32>) {
+    /// Gets the inner value with dependants
+    pub fn destruct(self) -> (T, Vec<u64>) {
         (self.value, self.dependents)
     }
     ///Constructs a Simple from a MathExpr
@@ -265,7 +270,7 @@ impl<T: Simplify> Simple<T> {
     }
     /// Construct a Simple without actually checking
     /// that it's simplified.
-    pub(super) const fn new_unchecked(value: T, dependants: Vec<u32>) -> Self {
+    pub(super) const fn new_unchecked(value: T, dependants: Vec<u64>) -> Self {
         Self {
             value,
             dependents: dependants,
